# perf: Batch calls to `ModuleLoader::prepare_load` for dynamic imports